+++ packages/implementations/react/src/composite/ColorPicker/src/components/Picker/index.ts
@@ -1,0 +1,4 @@
+import { Picker } from "./Picker";
+
+export { Picker } from "./Picker";
+export default Picker;